# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/bookmark":{"get":{"tags":["bookmark-resource"],"operationId":"list","responses":{"200":{"description":"OK","content":{"application/vnd.net.pbforge.bookmark-list-v1":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookmarkV1"}}}}}}},"put":{"tags":["bookmark-resource"],"operationId":"create","requestBody":{"content":{"application/vnd.net.pbforge.bookmark-creation-v1":{"schema":{"$ref":"#/components/schemas/BookmarkCreationV1"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/vnd.net.pbforge.bookmark-v1":{"schema":{"$ref":"#/components/schemas/BookmarkV1"}}}}}},"post":{"tags":["bookmark-resource"],"operationId":"update","requestBody":{"content":{"application/vnd.net.pbforge.bookmark-v1":{"schema":{"$ref":"#/components/schemas/BookmarkV1"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/vnd.net.pbforge.bookmark-v1":{"schema":{"$ref":"#/components/schemas/BookmarkV1"}}}}}}},"/bookmark/{id}":{"get":{"tags":["bookmark-resource"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.net.pbforge.bookmark-v1":{"schema":{"$ref":"#/components/schemas/BookmarkV1"}}}}}},"delete":{"tags":["bookmark-resource"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"BookmarkCreationV1":{"required":["name","url"],"type":"object","properties":{"name":{"type":"string","description":"The name of the bookmark"},"url":{"type":"string","description":"The URL of the bookmark"},"description":{"type":"string","description":"The description provided for a bookmark"}}},"BookmarkV1":{"required":["id","name","url"],"type":"object","properties":{"id":{"type":"string","description":"The id of the bookmark"},"name":{"type":"string","description":"The name of the bookmark"},"url":{"type":"string","description":"The URL of the bookmark"},"description":{"type":"string","description":"The description provided for a bookmark"}}}}}}